# openapi/ebmgr.yaml
openapi: 3.0.3
info:
  title: ebook manager (ebmgr) App
  description: ebook manager API
  version: 0.1.0
servers:
- url: http://localhost:3333/api/v1
  description: Local server
paths:
  /directories:
    get:
      x-swagger-router-controller: Default
      description: Returns all directories list
      operationId: getDirectories
      responses:
        "200":
          description: A list of directories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryEntry'
  /books:
    get:
      x-swagger-router-controller: Default
      description: Returns all books list
      operationId: getBooks
      responses:
        "200":
          description: A list of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
#                x-content-type: application/json
  /books/{vpath}:
    get:
      x-swagger-router-controller: Default
      description: Returns a book information
      operationId: getBook
      parameters:
        - in: path
          name: vpath
          schema:
            type: string
          required: true
          description: virtual path of the book
      responses:
        "404":
          description: A book with the given vpath was not found
        "403":
          description: A book exists but cannot get information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: book information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /books/{vpath}/thumbnail:
    get:
      x-swagger-router-controller: Default
      description: Returns a thumbnail of the book
      operationId: getBookThumbnail
      parameters:
        - in: path
          name: vpath
          schema:
            type: string
          required: true
          description: virtual path of the book
      responses:
        "404":
          description: A book with the given vpath was not found
        "403":
          description: A book exists but cannot get thumbnail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: A thumbnail of book
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
  /books/{vpath}/star:
    put:
      x-swagger-router-controller: Default
      description: add/remove star to the book
      operationId: setStar
      parameters:
        - in: path
          name: vpath
          schema:
            type: string
          required: true
          description: virtual path of the book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                star:
                  type: boolean
                  description: true if the book has star
                  example: true
      responses:
        "404":
          description: A book with the given vpath was not found
        "400":
          description: invalid parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "200":
          description: succeed
components:
  schemas:
    DirectoryEntry:
      type: object
      additionalProperties:
        type: object
      example:
        foo:
          bar: {}
        baz: {}
    Book:
      type: object
      properties:
        title:
          type: string
          description: title of book
          example: some interesting title
        vpath:
          type: string
          description: virtualized path of book
          example: /foo/bar/baz
      example:
        title: some interesting title
        path: /foo/bar/baz
    Error:
      type: object
      properties:
        message:
          type: string
          description: error message
          example: some error occured
  securitySchemes:
    bearerAuth:
      type: http
      description: API Key
      scheme: bearer